Resolved Slither divide-before-multiply warning

+++ contracts/contracts/strategies/NativeStaking/ValidatorAccountant.sol
@@ -116,7 +116,8 @@ abstract contract ValidatorAccountant is ValidatorRegistrator {
 
         // send the ETH that is from fully withdrawn validators to the Vault
         if (newSweptETH >= MAX_STAKE) {
-            uint256 fullyWithdrawnValidators = newSweptETH / MAX_STAKE;
+            // explicitly cast to uint256 as we want to round to a whole number of validators
+            uint256 fullyWithdrawnValidators = uint256(newSweptETH / MAX_STAKE);
             if (activeDepositedValidators < fullyWithdrawnValidators) {
                 return _failAccounting(pauseOnFail);
             }
